Change naming from "FunctionExecutor" to "Function"
Change naming from "FunctionExecutor" to "Function"
