+++ custom_components/extended_openai_conversation/config_flow.py
@@ -41,7 +41,7 @@ from .const import (
     CONF_CHAT_MODEL,
     CONF_CONTEXT_THRESHOLD,
     CONF_CONTEXT_TRUNCATE_STRATEGY,
-    CONF_FUNCTIONS,
+    CONF_FUNCTION_TOOLS,
     CONF_MAX_FUNCTION_CALLS_PER_CONVERSATION,
     CONF_MAX_TOKENS,
     CONF_ORGANIZATION,
@@ -60,7 +60,7 @@ from .const import (
     DEFAULT_API_PROVIDER,
     DEFAULT_CHAT_MODEL,
     DEFAULT_CONF_BASE_URL,
-    DEFAULT_CONF_FUNCTIONS,
+    DEFAULT_CONF_FUNCTION_TOOLS,
     DEFAULT_CONTEXT_THRESHOLD,
     DEFAULT_CONTEXT_TRUNCATE_STRATEGY,
     DEFAULT_CONVERSATION_NAME,
@@ -107,7 +107,9 @@ STEP_USER_DATA_SCHEMA = vol.Schema(
     }
 )
 
-DEFAULT_CONF_FUNCTIONS_STR = yaml.dump(DEFAULT_CONF_FUNCTIONS, sort_keys=False)
+DEFAULT_CONF_FUNCTION_TOOLS_STR = yaml.dump(
+    DEFAULT_CONF_FUNCTION_TOOLS, sort_keys=False
+)
 
 DEFAULT_OPTIONS = types.MappingProxyType(
     {
@@ -117,7 +119,7 @@ DEFAULT_OPTIONS = types.MappingProxyType(
         CONF_MAX_FUNCTION_CALLS_PER_CONVERSATION: DEFAULT_MAX_FUNCTION_CALLS_PER_CONVERSATION,
         CONF_TOP_P: DEFAULT_TOP_P,
         CONF_TEMPERATURE: DEFAULT_TEMPERATURE,
-        CONF_FUNCTIONS: DEFAULT_CONF_FUNCTIONS_STR,
+        CONF_FUNCTION_TOOLS: DEFAULT_CONF_FUNCTION_TOOLS_STR,
         CONF_CONTEXT_THRESHOLD: DEFAULT_CONTEXT_THRESHOLD,
         CONF_CONTEXT_TRUNCATE_STRATEGY: DEFAULT_CONTEXT_TRUNCATE_STRATEGY,
         CONF_SHORTEN_TOOL_CALL_ID: DEFAULT_SHORTEN_TOOL_CALL_ID,
@@ -440,8 +442,8 @@ class ExtendedOpenAISubentryFlowHandler(ConfigSubentryFlow):
                 )
             ),
             vol.Optional(
-                CONF_FUNCTIONS,
-                default=DEFAULT_CONF_FUNCTIONS_STR,
+                CONF_FUNCTION_TOOLS,
+                default=DEFAULT_CONF_FUNCTION_TOOLS_STR,
             ): TemplateSelector(),
             vol.Optional(
                 CONF_CONTEXT_THRESHOLD,
