+++ custom_components/extended_openai_conversation/functions/web.py
@@ -1,0 +1,189 @@
+"""Web functions for HTTP API calls and HTML scraping."""
+
+from __future__ import annotations
+
+import logging
+from typing import Any
+
+from bs4 import BeautifulSoup
+import voluptuous as vol
+
+from homeassistant.components import rest, scrape
+from homeassistant.const import (
+    CONF_ATTRIBUTE,
+    CONF_METHOD,
+    CONF_NAME,
+    CONF_PAYLOAD,
+    CONF_RESOURCE,
+    CONF_RESOURCE_TEMPLATE,
+    CONF_TIMEOUT,
+    CONF_VALUE_TEMPLATE,
+    CONF_VERIFY_SSL,
+)
+from homeassistant.core import HomeAssistant
+from homeassistant.helpers import config_validation as cv, llm
+from homeassistant.helpers.template import Template
+
+from ..const import CONF_PAYLOAD_TEMPLATE
+from .base import Function
+
+_LOGGER = logging.getLogger(__name__)
+
+
+def get_rest_data(
+    hass: HomeAssistant, rest_config: dict[str, Any], arguments: dict[str, Any]
+) -> rest.data.RestData:
+    """Create RestData from config with template rendering."""
+    rest_config.setdefault(CONF_METHOD, rest.const.DEFAULT_METHOD)
+    rest_config.setdefault(CONF_VERIFY_SSL, rest.const.DEFAULT_VERIFY_SSL)
+    rest_config.setdefault(CONF_TIMEOUT, rest.data.DEFAULT_TIMEOUT)
+    rest_config.setdefault(rest.const.CONF_ENCODING, rest.const.DEFAULT_ENCODING)
+
+    resource_template: Template | None = rest_config.get(CONF_RESOURCE_TEMPLATE)
+    if resource_template is not None:
+        rest_config.pop(CONF_RESOURCE_TEMPLATE)
+        rest_config[CONF_RESOURCE] = resource_template.async_render(
+            arguments, parse_result=False
+        )
+
+    payload_template: Template | None = rest_config.get(CONF_PAYLOAD_TEMPLATE)
+    if payload_template is not None:
+        rest_config.pop(CONF_PAYLOAD_TEMPLATE)
+        rest_config[CONF_PAYLOAD] = payload_template.async_render(
+            arguments, parse_result=False
+        )
+
+    return rest.create_rest_data_from_config(hass, rest_config)
+
+
+class RestFunction(Function):
+    """REST tool for HTTP API calls."""
+
+    def __init__(self) -> None:
+        """Initialize Rest tool."""
+        super().__init__(
+            vol.Schema(rest.RESOURCE_SCHEMA).extend(
+                {
+                    vol.Optional("value_template"): cv.template,
+                    vol.Optional("payload_template"): cv.template,
+                }
+            )
+        )
+
+    async def execute(
+        self,
+        hass: HomeAssistant,
+        function_config: dict[str, Any],
+        arguments: dict[str, Any],
+        llm_context: llm.LLMContext | None,
+        exposed_entities: list[dict[str, Any]],
+    ) -> Any:
+        """Execute REST API call."""
+        rest_data = get_rest_data(hass, function_config, arguments)
+
+        await rest_data.async_update()
+        value = rest_data.data_without_xml()
+        value_template = function_config.get(CONF_VALUE_TEMPLATE)
+
+        if value is not None and value_template is not None:
+            value = value_template.async_render_with_possible_json_value(
+                value, None, arguments
+            )
+
+        return value
+
+
+class ScrapeFunction(Function):
+    """Scrape tool for HTML content extraction."""
+
+    def __init__(self) -> None:
+        """Initialize Scrape tool."""
+        super().__init__(
+            scrape.COMBINED_SCHEMA.extend(
+                {
+                    vol.Optional("value_template"): cv.template,
+                    vol.Optional("payload_template"): cv.template,
+                }
+            )
+        )
+
+    async def execute(
+        self,
+        hass: HomeAssistant,
+        function_config: dict[str, Any],
+        arguments: dict[str, Any],
+        llm_context: llm.LLMContext | None,
+        exposed_entities: list[dict[str, Any]],
+    ) -> Any:
+        """Execute web scraping."""
+        rest_data = get_rest_data(hass, function_config, arguments)
+        coordinator = scrape.coordinator.ScrapeCoordinator(
+            hass,
+            None,
+            rest_data,
+            function_config,
+            scrape.const.DEFAULT_SCAN_INTERVAL,
+        )
+        await coordinator.async_refresh()
+
+        new_arguments = dict(arguments)
+
+        for sensor_config in function_config["sensor"]:
+            name: Template = sensor_config.get(CONF_NAME)
+            value = self._async_update_from_rest_data(
+                coordinator.data, sensor_config, arguments
+            )
+            new_arguments["value"] = value
+            if name:
+                new_arguments[name.async_render()] = value
+
+        result = new_arguments["value"]
+        value_template = function_config.get(CONF_VALUE_TEMPLATE)
+
+        if value_template is not None:
+            result = value_template.async_render_with_possible_json_value(
+                result, None, new_arguments
+            )
+
+        return result
+
+    def _async_update_from_rest_data(
+        self,
+        data: BeautifulSoup,
+        sensor_config: dict[str, Any],
+        arguments: dict[str, Any],
+    ) -> Any:
+        """Update state from the rest data."""
+        value = self._extract_value(data, sensor_config)
+        value_template = sensor_config.get(CONF_VALUE_TEMPLATE)
+
+        if value_template is not None:
+            value = value_template.async_render_with_possible_json_value(
+                value, None, arguments
+            )
+
+        return value
+
+    def _extract_value(self, data: BeautifulSoup, sensor_config: dict[str, Any]) -> Any:
+        """Parse the html extraction in the executor."""
+        value: str | list[str] | None
+        select = sensor_config[scrape.const.CONF_SELECT]
+        index = sensor_config.get(scrape.const.CONF_INDEX, 0)
+        attr = sensor_config.get(CONF_ATTRIBUTE)
+        try:
+            if attr is not None:
+                value = data.select(select)[index][attr]
+            else:
+                tag = data.select(select)[index]
+                if tag.name in ("style", "script", "template"):
+                    value = tag.string
+                else:
+                    value = tag.text
+        except IndexError:
+            _LOGGER.warning("Index '%s' not found", index)
+            value = None
+        except KeyError:
+            _LOGGER.warning("Attribute '%s' not found", attr)
+            value = None
+        _LOGGER.debug("Parsed value: %s", value)
+        return value
