--- tests/function_executors/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for FunctionExecutor classes in extended_openai_conversation."""
