+++ tests/functions/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for Tool classes in extended_openai_conversation."""
