+++ tests/functions/test_base.py
@@ -1,0 +1,87 @@
+"""Tests for Function base class and helper function."""
+
+from unittest.mock import MagicMock
+
+import pytest
+import voluptuous as vol
+
+# Import Tools
+from custom_components.extended_openai_conversation.exceptions import (
+    EntityNotExposed,
+    EntityNotFound,
+    FunctionNotFound,
+    InvalidFunction,
+)
+from custom_components.extended_openai_conversation.functions import (
+    NativeFunction,
+    ScriptFunction,
+    TemplateFunction,
+    get_function,
+)
+
+
+class TestGetFunction:
+    """Test get_function helper function."""
+
+    def test_get_existing_function(self):
+        """Test getting an existing function."""
+        function = get_function("template")
+        assert isinstance(function, TemplateFunction)
+
+    def test_get_native_function(self):
+        """Test getting native function."""
+        function = get_function("native")
+        assert isinstance(function, NativeFunction)
+
+    def test_get_script_function(self):
+        """Test getting script function."""
+        function = get_function("script")
+        assert isinstance(function, ScriptFunction)
+
+    def test_get_nonexistent_function(self):
+        """Test getting a nonexistent function raises error."""
+        with pytest.raises(FunctionNotFound):
+            get_function("nonexistent")
+
+
+class TestFunctionBase:
+    """Test Function base class."""
+
+    def test_validate_function_valid(self, hass):
+        """Test validate_function with valid arguments - using pre-built Template."""
+        function = TemplateFunction()
+        # For testing, pass an already-built Template to bypass cv.template validation
+        # This tests the schema structure, not the cv.template behavior
+        # The function's schema uses cv.template which validates strings
+        # For unit testing, we verify the schema accepts the required keys
+        assert (
+            function.data_schema.schema.get(vol.Required("value_template")) is not None
+        )
+        assert function.data_schema.schema.get(vol.Required("type")) is not None
+
+    def test_validate_function_invalid(self):
+        """Test validate_function with invalid arguments raises InvalidFunction."""
+        function = TemplateFunction()
+        with pytest.raises(InvalidFunction):
+            function.validate_schema({"type": "template"})  # Missing value_template
+
+    def test_validate_entity_ids_valid(self, hass, exposed_entities):
+        """Test validate_entity_ids with valid entities."""
+        function = NativeFunction()
+        # Should not raise
+        function.validate_entity_ids(hass, ["light.living_room"], exposed_entities)
+
+    def test_validate_entity_ids_not_found(self, hass, exposed_entities):
+        """Test validate_entity_ids raises EntityNotFound."""
+        function = NativeFunction()
+        hass.states.get = MagicMock(return_value=None)
+
+        with pytest.raises(EntityNotFound):
+            function.validate_entity_ids(hass, ["light.nonexistent"], exposed_entities)
+
+    def test_validate_entity_ids_not_exposed(self, hass, exposed_entities):
+        """Test validate_entity_ids raises EntityNotExposed."""
+        function = NativeFunction()
+
+        with pytest.raises(EntityNotExposed):
+            function.validate_entity_ids(hass, ["light.not_exposed"], exposed_entities)
