+++ tests/functions/test_scrape.py
@@ -1,42 +1,36 @@
-"""Tests for ScrapeFunctionExecutor using yaml definitions."""
+"""Tests for ScrapeFunction using yaml definitions."""
 
 from unittest.mock import AsyncMock, patch
 
 import pytest
 
-# Import FunctionExecutors and test helpers
-from custom_components.extended_openai_conversation.helpers import (
-    ScrapeFunctionExecutor,
-    get_function_executor,
-)
-from tests.helpers import get_function_from_yaml
+# Import Tools and test helpers
+from custom_components.extended_openai_conversation.functions import ScrapeFunction
+from tests.helpers import prepare_function_tool_from_yaml
 
 
-class TestScrapeFunctionExecutorYaml:
-    """Test ScrapeFunctionExecutor using yaml definitions."""
+class TestScrapeFunctionYaml:
+    """Test ScrapeFunction using yaml definitions."""
 
     @pytest.fixture
-    def executor(self):
-        """Create ScrapeFunctionExecutor instance."""
-        return ScrapeFunctionExecutor()
+    def function(self):
+        """Create ScrapeFunction instance."""
+        return ScrapeFunction()
 
     async def test_execute_scrape_from_yaml(
-        self, hass, executor, exposed_entities, llm_context
+        self, hass, function, exposed_entities, llm_context
     ):
         """Test web scraping from yaml definition."""
         # Load function from yaml
-        func_def = get_function_from_yaml("scrape_example.yaml")
-
-        # Process function through executor's to_arguments
-        function_executor = get_function_executor(func_def["function"]["type"])
-        processed_function = function_executor.to_arguments(func_def["function"])
+        function_tool = prepare_function_tool_from_yaml("scrape_example.yaml")
+        function_config = function_tool["function"]
 
         with (
             patch(
-                "custom_components.extended_openai_conversation.helpers.rest.create_rest_data_from_config"
+                "custom_components.extended_openai_conversation.functions.web.rest.create_rest_data_from_config"
             ) as mock_rest,
             patch(
-                "custom_components.extended_openai_conversation.helpers.scrape.coordinator.ScrapeCoordinator"
+                "custom_components.extended_openai_conversation.functions.web.scrape.coordinator.ScrapeCoordinator"
             ) as mock_coordinator_class,
         ):
             from bs4 import BeautifulSoup
@@ -69,8 +63,8 @@ class TestScrapeFunctionExecutorYaml:
             # Arguments based on yaml spec parameters (category is optional)
             arguments = {}
 
-            result = await executor.execute(
-                hass, processed_function, arguments, llm_context, exposed_entities
+            result = await function.execute(
+                hass, function_config, arguments, llm_context, exposed_entities
             )
 
             # Should return scraped data
